+++ src/main/kotlin/com/github/sleepypanda/feesh/features/alerts/AnyReindrakeAlert.kt
@@ -7,15 +7,19 @@ import com.github.sleepypanda.feesh.utils.SoundUtils
 import com.github.sleepypanda.feesh.utils.RegisterUtils
 import com.github.sleepypanda.feesh.utils.WorldUtils
 import com.github.sleepypanda.feesh.utils.ChatUtils
+import com.github.sleepypanda.feesh.utils.PlayerUtils
+import net.minecraft.text.Text
 
 object AnyReindrakeAlert {
     // WOAH! [MVP+] MoonTheSadFisher summoned a Reindrake from the depths!
     // WOAH! [MVP+] MoonTheSadFisher summoned TWO Reindrakes from the depths!
-    const val REINDRAKE_PATTERN = "^WOAH! (?<playerNameAndRank>.*) summoned (a|TWO|two) Reindrake(s)? from the depths!$"
+    // §c§lWOAH! §6[MVP§0++§6] Dulkir§f §csummoned a §4Reindrake §cfrom the depths!
+    // §c§lWOAH! §b[MVP§d+§b] MoonTheSadFisher§f §csummoned a §4Reindrake §cfrom the depths!
+    const val REINDRAKE_PATTERN = "^§c§lWOAH! (?<playerNameAndRank>.*) §csummoned (a|TWO|two) §4Reindrake(s)? §cfrom the depths!$"
     val reindrake = SeaCreatures.allSeaCreatures.find { it.name == "Reindrake" }!!
 
     fun init() {
-        RegisterUtils.chat(Regex(REINDRAKE_PATTERN)) { _, matchResult -> onAnyReindrake(reindrake.boldDisplayName, reindrake.rarityColorCode, matchResult) }
+        RegisterUtils.chat(Regex(REINDRAKE_PATTERN), noFormatting = false) { _, matchResult -> onAnyReindrake(reindrake.boldDisplayName, reindrake.rarityColorCode, matchResult) }
     }
 
     private fun onAnyReindrake(boldDisplayName: String, rarityColorCode: String, matchResult: MatchResult) {
@@ -23,7 +27,10 @@ object AnyReindrakeAlert {
         if (!WorldUtils.isInSkyblock() || !Alerts.alertOnAnyReindrake || WorldUtils.getWorldName() != WorldUtils.JERRY_WORKSHOP) return
 
         val isDoubleHook = matchResult.groupValues[2].equals("two", ignoreCase = true)
-        CommonUtils.showTitle(SeaCreatures.getTitle(reindrake.name, isDoubleHook))
+        val playerNameAndRank = matchResult.groups.get("playerNameAndRank")?.value ?: ""
+        val name = PlayerUtils.getFormattedPlayerNameFromPartyChat(playerNameAndRank) ?: ""
+
+        CommonUtils.showTitle(SeaCreatures.getTitle(reindrake.name, isDoubleHook), name)
         ChatUtils.sendLocalChatWithCommand("Click to warp to Jerry's Workshop spawn point!", "warp jerry", true)
         SoundUtils.playSound()
     }